+++ src/Model/Response.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Supermarket;
+namespace Supermarket\Model;
 
 class Response
 {
